Refactoring Encoder/Decoder.
