+++ websocket-commons/src/main/java/org/lfenergy/compas/core/websocket/WebsocketHandler.java
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Alliander N.V.
 //
 // SPDX-License-Identifier: Apache-2.0
-package org.lfenergy.compas.core.websocket.handler;
+package org.lfenergy.compas.core.websocket;
 
 import org.lfenergy.compas.core.commons.exception.CompasException;
 import org.lfenergy.compas.core.commons.model.ErrorResponse;
