+++ websocket-commons/src/test/java/org/lfenergy/compas/core/websocket/WebsocketHandlerTest.java
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Alliander N.V.
 //
 // SPDX-License-Identifier: Apache-2.0
-package org.lfenergy.compas.core.websocket.handler;
+package org.lfenergy.compas.core.websocket;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
